refactor: move prefill scheduling to dedicated module
